package.yaml: Specify a lower version bound for the websocket library version
Saw an instance where circleci picked up a lower incompatible version of
websocket library which caused the build to fail.

This commit also adds a ghc flag that we incorrectly added to the cabal file
where it should have been added to the package.yaml file as cabal file is
generated by hpack.